Add map class, movement, shortcut class, art class, package cleanup
The map can be created in the new Map class which hold anything that is useful for the map to know, and it allows any size, currently is only rectangular, something that is likely to stay this way. Added movement of a creature, possibly will use in the future for monsters as well not only the hero, should work properly. Two new classes are added, one to keep track of all the shortcuts for the moement options, this is so that we can adjust the shortcuts at will and allow the player to do it from in game. The other is a class for art, which will hold any possible multi-line ASCII art that we don't want to clutter our code, everything is supposed to be used from there if it's a multi-line printing. Some packages were added to make it more clear while working on the code about what suits where.