Exit game. Music
One room is marked as the exit. When the hero is over the exit an appropriate message is shown and a new option is shown in the menu options where the hero can exit the game and keep his high score. The hero can chose not to of course and can go back later after he has collected more loot. For music, it has two tracks at the moment, that are played randomly when in the main menu, but during a fight a special track is played until the fight is over and then one of the other tracks continues playing until the next fight or the game is over.